Cleaning up some things
