+++ BlockRegeneratorPlugin/src/main/resources/plugin.yml
@@ -1,0 +1,5 @@
+name: BlockRegenerator
+version: 1.0
+main: me.joshuaemq.blockregenerator.BlockRegenerator
+description: RegenerateBlocks!
+depend: [WorldGuard, WorldEdit]
