--- BlockRegeneratorPlugin/src/plugin.yml
@@ -1,4 +1,0 @@
-name: BlockRegenerator
-version: 1.0
-main: me.joshuaemq.blockRegenerator.BlockRegenerator
-description: RegenerateBlocks!
